Remove TransactionListFragment (#503)
* Remove TransactionListFragment
* Simplify MainActivity layout
* Address PR review
